DeblurGANv2/requirements.txt
torchsummary~=1.5.1
pretrainedmodels~=0.7.4
numpy~=1.26.4
opencv-python-headless
joblib~=1.4.2
albumentations>=1.0.0
scikit-image~=0.24.0
tqdm~=4.66.5
glog~=0.3.1
tensorboardx~=2.6.2.2
fire~=0.6.0
torch~=2.3.0+cu118
torchvision~=0.18.0+cu118
opencv-python~=4.10.0.84
pyyaml~=6.0.2
pillow~=10.3.0